refactor: gitignore
