--- src/medidas_geneticas.h
@@ -1,4 +1,0 @@
-#include "./libs/types.h"
-
-void densityPopulation(populacao *populations,int island_size);
-void densityWorld(populacao *populations,int island_size);
